+++ Gopkg.toml
@@ -44,7 +44,7 @@
   name = "github.com/solo-io/go-checkpoint"
   version = "0.0.4"
 
-[[constraint]]
+[[override]]
   name = "github.com/solo-io/solo-kit"
   version = "0.6.1"
